requirements.txt
crewai
crewai-tools
python-dotenv
requests
beautifulsoup4
langchain
langchain-groq
langsmith